bms-lambda/src/requirements.txt
# Database
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# JSON handling
orjson==3.9.10

# Date/time utilities
python-dateutil==2.8.2

# Logging
structlog==23.2.0

# Validation
pydantic==2.5.0

# Testing (optional)
pytest==7.4.3
pytest-mock==3.12.0

# Development (optional)
black==23.11.0
flake8==6.1.0
mypy==1.7.1